+++ backend/src/main/java/ru/laspace/backend/service/Pr01Service.java
@@ -1,5 +1,5 @@
 package ru.laspace.backend.service;
 
 public interface Pr01Service {
-    String generateAndSave(Long numRp, Long numKa);
+    String generateAndSave(Integer numRp, Integer numKa);
 }
