+++ backend/src/main/java/ru/laspace/backend/service/Pr03Service.java
@@ -1,5 +1,5 @@
 package ru.laspace.backend.service;
 
 public interface Pr03Service {
-    String generateAndSave(Long numRp, Long numKa);
+    String generateAndSave(Integer numRp, Integer numKa);
 }
