+++ backend/src/main/java/ru/laspace/backend/service/ProgramNumberService.java
@@ -1,5 +1,5 @@
 package ru.laspace.backend.service;
 
 public interface ProgramNumberService {
-    Long generateNextProgramNumber(Long numKa);
+    Integer generateNextProgramNumber(Integer numKa);
 }
